Created the SuccessPanel for when the password reset link is successfully sent
